Feature/rating service kafka

+++ rating-service/src/main/java/by/shestakov/ratingservice/dto/response/AverageRatingResponse.java
@@ -1,6 +1,8 @@
 package by.shestakov.ratingservice.dto.response;
 
+import java.math.BigDecimal;
+
 public record AverageRatingResponse(
-        Double average
+        BigDecimal average
 ) {
 }
